backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
alembic==1.13.3
PyMySQL==1.1.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
cryptography==43.0.1
